ajustando atualizacao automatica e retorno API

+++ RestAPI/Controllers/LivroController.cs
@@ -25,7 +25,12 @@ namespace RestAPI.Controllers
         [Produces(applicationJson)]
         public IActionResult FindAll()
         {
-            return Ok(this.livroService.FindAll());
+            IList<LivroDTO> livros = this.livroService.FindAll();
+
+            if (livros == null || livros.Count == 0)
+                return NotFound();
+
+            return Ok(livros);
         }
 
         [HttpGet("{id}")]
@@ -47,6 +52,10 @@ namespace RestAPI.Controllers
         [Consumes(applicationJson)]
         public IActionResult Save(LivroDTO livro)
         {
+            IActionResult validacao = this.ValidarLivro(livro);
+
+            if (validacao != null) return validacao;
+
             livro = this.livroService.Save(livro);
 
             return CreatedAtAction(nameof(Find), new { id = livro.Id }, livro);
@@ -61,6 +70,10 @@ namespace RestAPI.Controllers
             if (this.livroService.Find(id) == null)
                 return NotFound();
 
+            IActionResult validacao = this.ValidarLivro(livro);
+
+            if (validacao != null) return validacao;
+
             livro.Id = id;
 
             return Accepted(this.livroService.Save(livro));
@@ -70,6 +83,9 @@ namespace RestAPI.Controllers
         [Authorize]
         public IActionResult Delete(Int64 id)
         {
+            if (this.livroService.Find(id) == null)
+                return NotFound();
+
             this.livroService.Delete(id);
 
             return NoContent();
@@ -80,7 +96,22 @@ namespace RestAPI.Controllers
         [Produces(applicationJson)]
         public IActionResult FindByTitle(String titulo)
         {
-            return Ok(this.livroService.FindByTitle(titulo));
+            IList<LivroDTO> livros = this.livroService.FindByTitle(titulo);
+
+            if (livros == null || livros.Count == 0)
+                return NotFound();
+
+            return Ok(livros);
+        }
+
+        private IActionResult ValidarLivro(LivroDTO livro)
+        {
+            String validacao = this.livroService.ValidarLivro(livro);
+
+            if (!String.IsNullOrEmpty(validacao))
+                return BadRequest(validacao);
+
+            return null;
         }
     }
 }
